Factor out getPeerDependencies and remove from public API.
